Arquitetura do backend refatorada
Backend agora está organizado seguindo o padrão de Arquitetura em Camadas (Layered Architecture)

Pasta "database" renomeada para "backend"
