+++ backend/requirements.txt
@@ -1,0 +1,6 @@
+flask
+flask-cors
+mysql-connector-python
+bcrypt
+PyJWT
+python-dotenv
